[plugin-framework-foundation] Address review: break circular import, mut reg, fix stale comment
- Extract types + `createRegistry` + `PluginRegisterError` from `src/plugins/index.ts` into a new `src/plugins/registry.ts`. `index.ts` becomes a thin barrel re-exporting from `./registry`, `./registry-provider`, and `./use-registry`. The internal provider / context / hook files now import from `./registry` directly, breaking the `index ↔ registry-provider` circular dependency that could destabilise HMR / Fast Refresh.

- `plugins::build_registry()` now declares `let mut reg` (with `#[allow(unused_mut)]` for today's empty body) so the commented `reg.register_*(...)` extension points compile the instant sub-issues #96/#97/#98 uncomment them.

- Update the stale header comment in `src/plugins/use-registry.ts` (and clarify `registry-context.ts`) to match the actual three-file split (context constant + provider component + hook), instead of saying "the context file only exports a component".
