+++ src/plugins/use-registry.ts
@@ -1,11 +1,14 @@
-// `useRegistry` hook split from `registry-context.tsx` so the context
-// file only exports a component (keeps Fast Refresh happy under the
-// `react-refresh/only-export-components` lint rule).
+// `useRegistry` hook for the plugin registry context (issue #95).
+//
+// Kept in its own file so the provider component (`./registry-provider.tsx`)
+// and the context constant (`./registry-context.ts`) each export only one
+// concern apiece — that's what keeps every file in this trio Fast-Refresh
+// friendly under the `react-refresh/only-export-components` lint rule.
 
 import { useContext } from 'react';
 
 import { PluginRegistryContext } from './registry-context';
-import type { PluginRegistry } from './index';
+import type { PluginRegistry } from './registry';
 
 /**
  * Access the registry from any component beneath
